+++ khepera/evoSlave/trunk/genes.h
@@ -1,8 +1,8 @@
 /*****************************************************************/
-/*  genes.h
-/*	Created by richard at 21:20 on 17/11/2010
-/*	Copyright Durham University 2010, all rights reserved
-/*
+//  genes.h
+//	Created by richard at 21:20 on 17/11/2010
+//	Copyright Durham University 2010, all rights reserved
+//
 /*****************************************************************/
 #ifndef INC_GENES
 #define INC_GENES
@@ -13,8 +13,8 @@
 
 
 //Genotype stored as global:
-extern int nInputs, nHiddens, nOutputs;
-extern int *weights;
+extern unsigned short int nInputs, nHiddens, nOutputs;
+extern float *weights;
 
 void readGenotype(const char *filename);
 void printGenotype(void);
